Add support for signature tags
* The idea here is we'd like to get a webhook notification when the
signatures notes ref changes. Unfortunately, github does not fire
webhooks for refs outside refs/heads or refs/tags. To get webhooks to
fire for these events, we just create a tag which moves each time the
signatures notes ref is updated, and push that tag, which will fire off
a webhook. Note that the actual commit that the tag points to is of
absolutely no importance and one should not rely on the latest-signature
tag to actually point at the latest signature, any notes ref, or really
anything of use.

Signed-off-by: Tyler Levine <tyler@bitgo.com>